+++ Sources/Core/ref.swift
@@ -5,7 +5,7 @@ func ref(context: Context, reference: Any, instance: Any, schema: [String: Any])
 
   guard let document = context.resolve(ref: reference) else {
     return AnySequence([
-      ValidationError("Reference not found '\(reference)'"),
+      ValidationError("Reference not found '\(reference)'", instanceLocation: nil),
     ])
   }