+++ Tests/JSONSchemaTests/JSONSchemaTests.swift
@@ -29,14 +29,14 @@ class JSONSchemaTests: XCTestCase {
   }
 
   func testSuccessfulValidation() {
-    XCTAssertTrue(schema.validate([String: Any]()).valid)
+    try XCTAssertTrue(schema.validate([String: Any]()).valid)
   }
 
   func testUnsuccessfulValidation() {
-    XCTAssertFalse(schema.validate([String]()).valid)
+    try XCTAssertFalse(schema.validate([String]()).valid)
   }
 
-  func testReadme() {
+  func testReadme() throws {
     let schema = Schema([
       "type": "object",
       "properties": [
@@ -46,11 +46,11 @@ class JSONSchemaTests: XCTestCase {
       "required": ["name"],
     ])
 
-    XCTAssertTrue(schema.validate(["name": "Eggs", "price": 34.99]).valid)
-    XCTAssertFalse(schema.validate(["price": 34.99]).valid)
+    try XCTAssertTrue(schema.validate(["name": "Eggs", "price": 34.99]).valid)
+    try XCTAssertFalse(schema.validate(["price": 34.99]).valid)
   }
 
-  func testIterableInterface() {
+  func testIterableInterface() throws {
     let schema = Schema([
       "type": "object",
       "properties": [
@@ -61,14 +61,14 @@ class JSONSchemaTests: XCTestCase {
     ])
 
     var counter = 0
-    for error in schema.validate(["price": 34.99]) {
+    for error in try schema.validate(["price": 34.99]) {
       XCTAssertEqual(error.description, "Required property 'name' is missing")
       counter += 1
     }
 
     XCTAssertEqual(counter, 1)
 
-    let result = Array(schema.validate(["price": 34.99]))
+    let result = try Array(schema.validate(["price": 34.99]))
 
     XCTAssertEqual(result.map(\.description), [
       "Required property 'name' is missing"
@@ -89,8 +89,8 @@ class ValidateTests: XCTestCase {
       "required": ["name"],
     ]
 
-    XCTAssertTrue(validate(["name": "Eggs", "price": 34.99], schema: schema).valid)
-    XCTAssertFalse(validate(["price": 34.99], schema: schema).valid)
+    try XCTAssertTrue(validate(["name": "Eggs", "price": 34.99], schema: schema).valid)
+    try XCTAssertFalse(validate(["price": 34.99], schema: schema).valid)
   }
 
   func testValidateDraft6() {
@@ -104,8 +104,8 @@ class ValidateTests: XCTestCase {
       "required": ["name"],
     ]
 
-    XCTAssertTrue(validate(["name": "Eggs", "price": 34.99], schema: schema).valid)
-    XCTAssertFalse(validate(["price": 34.99], schema: schema).valid)
+    try XCTAssertTrue(validate(["name": "Eggs", "price": 34.99], schema: schema).valid)
+    try XCTAssertFalse(validate(["price": 34.99], schema: schema).valid)
   }
 
   func testDraft6ValidatorIsAvailable() {
@@ -113,7 +113,7 @@ class ValidateTests: XCTestCase {
     XCTAssertTrue(result is Draft6Validator, "Unexpected type of validator \(result)")
   }
 
-  func testValidateDraft7() {
+  func testValidateDraft7() throws {
     let schema: [String: Any]  = [
       "$schema": "http://json-schema.org/draft-07/schema#",
       "type": "object",
@@ -124,17 +124,17 @@ class ValidateTests: XCTestCase {
       "required": ["name"],
     ]
 
-    XCTAssertTrue(validate(["name": "Eggs", "price": 34.99], schema: schema).valid)
-    XCTAssertFalse(validate(["price": 34.99], schema: schema).valid)
+    try XCTAssertTrue(validate(["name": "Eggs", "price": 34.99], schema: schema).valid)
+    try XCTAssertFalse(validate(["price": 34.99], schema: schema).valid)
   }
 
-  func testValidatesRequired() {
+  func testValidatesRequired() throws {
     let schema: [String: Any] = [
       "$schema": "http://json-schema.org/draft-07/schema#",
       "required": ["one", "two", "three"],
     ]
 
-    let result = validate(["one": true, "three": true], schema: schema)
+    let result = try validate(["one": true, "three": true], schema: schema)
 
     switch result {
     case .valid:
@@ -146,7 +146,7 @@ class ValidateTests: XCTestCase {
     }
   }
 
-  func testRequiredValidationLocation() {
+  func testRequiredValidationLocation() throws {
     let schema: [String: Any] = [
       "$schema": "http://json-schema.org/draft-07/schema#",
       "items": [
@@ -154,7 +154,7 @@ class ValidateTests: XCTestCase {
       ]
     ]
 
-    let result = validate([[:]], schema: schema)
+    let result = try validate([[:]], schema: schema)
 
     switch result {
     case .valid: